+++ src/test/java/LoginTests.java
@@ -10,9 +10,7 @@ public class LoginTests extends BaseTest {
     public String url = "https://qa.koel.app/";
     @Test
     public void loginEmptyEmailPassword() {
-        // test for forking
-//      Added ChromeOptions argument below to fix websocket error
-        navigateToPage();
+        //navigateToPage();
         Assert.assertEquals(driver.getCurrentUrl(), url);
         driver.quit();
     }
