+++ src/test/resources/TestNG.xml
@@ -1,14 +1,18 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
 
-<suite name="TestNG">
+<suite name="homework18">
     <parameter name="baseUrl" value="https://qa.koel.app/"/>
-    <test name="Example Test" preserve-order="false">
+    <test name="homework18" preserve-order="false">
         <classes>
-<!--            <class name="LoginTests"/>-->
-            <class name="BaseTest"/>-->
+            <class name="LoginTests"/>
+            <class name="BaseTest"/>
 <!--            <class name="homework17"/>-->
-            <class name="homework18"/>
+            <class name="homework18">
+                <methods>
+                    <include name="playSong"/>
+                </methods>
+            </class>
 
         </classes>
     </test>
